user의 grade 업데이트

+++ src/api/review/entities/review.entity.ts
@@ -12,7 +12,7 @@ export class Reviews {
   @PrimaryGeneratedColumn({ type: "int", name: "id" })
   id: number;
 
-  @Column("int", { name: "grade" })
+  @Column("float", { name: "grade" })
   grade: number;
 
   @Column("varchar", { name: "review", length: 2000 })
